Cleaning the code...
